Refactored code from three files into data_collection.py FILE.
